Do not walk the filetree setting quotas
From the git logs, this seems to have been a workaround for a time when a version of compute created user volumes
outside of the purview of the fileservice or quota manager, which seems not to be the case any more. We expect each
volume is created with the cache, and this endpoint can still be called via the API with credentials. We may choose to
remove it after a period of observation where it is no longer deemed necessary.

It is expected that some pressure will be releived on the underlying filesystem with this change.